separate tests.report module (lua) disable all report testing for now
