+++ scripts/tests/e3/init.lua
@@ -6,5 +6,6 @@ require 'tests.e3.parser'
 require 'tests.e3.morale'
 require 'tests.orders'
 require 'tests.common'
+-- require 'tests.report'
 require 'tests.magicbag'
 require 'tests.process'
