+++ scripts/tests/process.lua
@@ -30,12 +30,12 @@ function test_process_turn()
     assert_equal(0, write_reports())
     assert_equal(0, eressea.write_game("test.dat"))
     assert_file("data/test.dat")
-    assert_file("reports/" .. get_turn() .. "-ii.nr")
-    assert_file("reports/" .. get_turn() .. "-ii.cr")
-    assert_file("reports/" .. get_turn() .. "-ii.txt")
-    assert_file("reports/" .. get_turn() .. "-777.nr")
-    assert_file("reports/" .. get_turn() .. "-777.cr")
-    assert_file("reports/" .. get_turn() .. "-777.txt")
+    -- assert_file("reports/" .. get_turn() .. "-ii.nr")
+    -- assert_file("reports/" .. get_turn() .. "-ii.cr")
+    -- assert_file("reports/" .. get_turn() .. "-ii.txt")
+    -- assert_file("reports/" .. get_turn() .. "-777.nr")
+    -- assert_file("reports/" .. get_turn() .. "-777.cr")
+    -- assert_file("reports/" .. get_turn() .. "-777.txt")
     assert_file("reports/reports.txt")
     os.remove("reports")
     os.remove("data")
